Add readonly mode for file-based pages and widgets

+++ bundles/org.openhab.ui/web/src/pages/settings/pages/pages-list.vue
@@ -134,6 +134,9 @@
                 </div>
               </template>
               <!-- <span class="item-initial">{{page.config.label[0].toUpperCase()}}</span> -->
+              <template #after-title>
+                <f7-icon v-if="page.editable === false" f7="lock_fill" size="1rem" color="gray" />
+              </template>
               <template #media>
                 <oh-icon
                   :color="page.config.sidebar || page.uid === 'overview' ? '' : 'gray'"
@@ -405,7 +408,18 @@ export default {
       })
     },
     doRemoveSelected() {
-      let dialog = f7.dialog.progress(`Deleting Pages...`)
+      // TODO
+      if (
+        this.selectedItems.some((p) => {
+          const uid = p.startsWith('system:sitemap:') ? p.replace('system:sitemap:', '') : p.replace('ui:page:', '')
+          return this.pages.find((page) => page.uid === uid)?.editable === false
+        })
+      ) {
+        f7.dialog.alert('Some of the selected pages are not modifiable because they have been provisioned by files')
+        return
+      }
+
+      let dialog = f7.dialog.progress(`Deleting ${this.showSitemaps ? 'Sitemaps' : 'Pages'}...`)
 
       const promises = this.selection.map((p) => {
         return this.$oh.api.delete('/rest/ui/components/ui:page/' + p)
